Updated power from Double to Int

+++ src/HashedOperation.hs
@@ -129,7 +129,7 @@ product es = Just . applyNary (naryET Mul ElementDefault) $ es
 -- | Element-wise multiplication
 --
 instance (DimensionType d, NumType et) => PowerOp (Expression d et) where
-    (^) :: Expression d et -> Double -> Expression d et
+    (^) :: Expression d et -> Int -> Expression d et
     (^) e1 x = applyUnary ((unary $ Power x) `hasShape` expressionShape e1) e1
 
 -- | Scale in vector space
